refactor: 컴포넌트를 도메인 별로 폴더 분리
